+++ CHANGELOG.md
@@ -10,9 +10,11 @@ or for custom notifications using *Generic* transport
 - *Target* definition for *Delivery* or *Transport* can now be more flexible
   - Allows a structured dict, a single value, or list of strings.
   - Structure only required where category can't be inferred, so entity_ids, device_ids, email and phone numbers are fine
-- *Deliver* has a new `target_definition` key, taking values of `fixed`,`default` or `merge`.
+- *Deliver* has a new `target_definition` key, taking values of `fixed`,`default`, `merge` or `merge_default`
   - `default` only uses the Delivery target if there's no target on the notification action call
-  - `merge` combines the targets in the Delivery with any on the action call
+  - `merge` combines the targets in the Delivery with any on the action call, only where delivery already has a target
+  - `merge_default` combines the targets in the Delivery with any on the action call, or if there's
+  no target on the notification, it defaults to the Delivery target
   - `fixed` only ever delivers to the targets in the Delivery config, ignoring any direct or indirect (for example `person_id`) in the action call
 ### Internal
 - *Transport* implementations simplified, `select_targets` now replaced by options to select